Remove config and refactor.
