+++ crates/notary/server/fixture/config/config.yaml
@@ -28,11 +28,13 @@ notarization:
   max_sent_data: 4096
   max_recv_data: 16384
   timeout: 1800
-  private_key_path: "../fixture/notary/notary.key"
-  public_key_path: "../fixture/notary/notary.pub"
+  private_key_path: "../notary/notary.key"
+  public_key_path: "../notary/notary.pub"
 
 tls:
   enabled: false
+  private_key_path: "../tls/key.pem"
+  certificate_path: "../tls/cert.pem"
 
 log:
   level: DEBUG
@@ -40,3 +42,4 @@ log:
 
 auth:
   enabled: false
+  whitelist_path: "../auth/whitelist.csv"
