+++ crates/notary/server/src/lib.rs
@@ -1,7 +1,6 @@
 mod auth;
 mod cli;
 mod config;
-mod api;
 mod error;
 mod middleware;
 mod server;
@@ -10,14 +9,16 @@ mod service;
 mod signing;
 #[cfg(feature = "tee_quote")]
 mod tee;
+mod types;
 mod util;
 
-pub use api::{ClientType, NotarizationSessionRequest, NotarizationSessionResponse};
 pub use cli::CliFields;
 pub use config::{
-    AuthorizationProperties, LogProperties, NotarizationProperties, NotaryServerProperties, TLSProperties,
+    AuthorizationProperties, LogProperties, NotarizationProperties, NotaryServerProperties,
+    TLSProperties,
 };
 pub use error::NotaryServerError;
 pub use server::{read_pem_file, run_server};
 pub use server_tracing::init_tracing;
+pub use types::{ClientType, NotarizationSessionRequest, NotarizationSessionResponse};
 pub use util::parse_config_file;
