+++ crates/notary/server/src/middleware.rs
@@ -3,11 +3,7 @@ use axum_core::extract::{FromRef, FromRequestParts};
 use std::collections::HashMap;
 use tracing::{error, trace};
 
-use crate::{
-    auth::AuthorizationWhitelistRecord,
-    api::NotaryGlobals,
-    NotaryServerError,
-};
+use crate::{auth::AuthorizationWhitelistRecord, types::NotaryGlobals, NotaryServerError};
 
 /// Auth middleware to prevent DOS
 pub struct AuthorizationMiddleware;
@@ -62,9 +58,7 @@ fn api_key_is_valid(
 #[cfg(test)]
 mod test {
     use super::{api_key_is_valid, HashMap};
-    use crate::auth::{
-        authorization_whitelist_vec_into_hashmap, AuthorizationWhitelistRecord,
-    };
+    use crate::auth::{authorization_whitelist_vec_into_hashmap, AuthorizationWhitelistRecord};
     use std::sync::Arc;
 
     fn get_whitelist_fixture() -> HashMap<String, AuthorizationWhitelistRecord> {
